requirements.txt
azure-mgmt-resource>=9.0.0
azure-identity>=1.5.0
azure-mgmt-resourcegraph>=2.0.0
azure-cli-core>=2.5.0
